+++ azure_terminal/core/main.py
@@ -1,2 +1,12 @@
-from azure_terminal.core.commands import *
+# Importing
+from azure_terminal.core.command_handling import execute_command
 
+# Main Function
+def azure_terminal_main():
+    while True:
+        prompt: str = str(input("azure-core$ ")).strip()
+        execute_command(prompt)
+
+# Calling main function
+if __name__ == "__main__":
+    azure_terminal_main()