+++ azure_terminal/modules/module1.py
@@ -1,4 +1,0 @@
-# module1.py in the "modules" folder
-
-def execute():
-    print("Module 1 command executed!")