+++ azure_terminal/modules/module2.py
@@ -1,4 +1,0 @@
-# module2.py in the "modules" folder
-
-def execute():
-    print("Module 2 command executed!")