+++ azure_terminal/docs/CHANGELOG.md
@@ -8,10 +8,15 @@ All notable changes to Azure Terminal will be documented in this file. This proj
 ### Added
 - Basic structure for Azure Terminal.
 
+## v0.0.1 - 27/2/2025 - Build #1
+### Added
+- Basic system commands like `ls` & `quit`.
+- Custom modules can now be added modularity.
+
 ---
 
 ### Note:
-- All changes, updates, and fixes will be reflected here to keep track of Azure Terminals's evolution.
+- All changes, updates, and fixes will be reflected here to keep track of Azure Terminal's evolution.
 - This changelog follows [Semantic Versioning](https://semver.org/) to indicate the types of changes made in each release.
 - All dates follow the `DD/MM/YYYY` format.
 - The build number increases with each version to indicate the release order